Convert atomInfo to int64_t
This ensures we can use bits from a 64-bit range if we ever need
that. Or easily switch to 32-bit range if that's advantageous.

Refs #4004

+++ api/nblib/forcecalculator.cpp
@@ -70,9 +70,9 @@ void ForceCalculator::compute(gmx::ArrayRef<const Vec3> coordinates, gmx::ArrayR
     gmxForceCalculator_->compute(coordinates, forces);
 }
 
-void ForceCalculator::updatePairList(gmx::ArrayRef<const int> particleInfoAllVdW,
-                                     gmx::ArrayRef<Vec3>      coordinates,
-                                     const Box&               box)
+void ForceCalculator::updatePairList(gmx::ArrayRef<const int64_t> particleInfoAllVdW,
+                                     gmx::ArrayRef<Vec3>          coordinates,
+                                     const Box&                   box)
 {
     gmxForceCalculator_->setParticlesOnGrid(particleInfoAllVdW, coordinates, box);
 }
