+++ api/nblib/gmxsetup.h
@@ -121,7 +121,7 @@ private:
     std::vector<real> nonbondedParameters_;
 
     //! Particle info where all particles are marked to have Van der Waals interactions
-    std::vector<int> particleInfoAllVdw_;
+    std::vector<int64_t> particleInfoAllVdw_;
 
     //! GROMACS force calculator to compute forces
     std::unique_ptr<GmxForceCalculator> gmxForceCalculator_;
