+++ src/gromacs/nbnxm/grid.cpp
@@ -798,12 +798,12 @@ static void print_bbsizes_supersub(FILE* fp, const Grid& grid)
  *
  * Sorts atoms on LJ coefficients: !=0 first, ==0 at the end.
  */
-static void sort_cluster_on_flag(int                      numAtomsInCluster,
-                                 int                      atomStart,
-                                 int                      atomEnd,
-                                 gmx::ArrayRef<const int> atomInfo,
-                                 gmx::ArrayRef<int>       order,
-                                 int*                     flags)
+static void sort_cluster_on_flag(int                          numAtomsInCluster,
+                                 int                          atomStart,
+                                 int                          atomEnd,
+                                 gmx::ArrayRef<const int64_t> atomInfo,
+                                 gmx::ArrayRef<int>           order,
+                                 int*                         flags)
 {
     constexpr int c_maxNumAtomsInCluster = 8;
     int           sort1[c_maxNumAtomsInCluster];
@@ -879,7 +879,7 @@ void Grid::fillCell(GridSetData*                   gridSetData,
                     nbnxn_atomdata_t*              nbat,
                     int                            atomStart,
                     int                            atomEnd,
-                    gmx::ArrayRef<const int>       atomInfo,
+                    gmx::ArrayRef<const int64_t>   atomInfo,
                     gmx::ArrayRef<const gmx::RVec> x,
                     BoundingBox gmx_unused* bb_work_aligned)
 {
@@ -1025,7 +1025,7 @@ void Grid::fillCell(GridSetData*                   gridSetData,
 
 void Grid::sortColumnsCpuGeometry(GridSetData*                   gridSetData,
                                   int                            dd_zone,
-                                  gmx::ArrayRef<const int>       atomInfo,
+                                  gmx::ArrayRef<const int64_t>   atomInfo,
                                   gmx::ArrayRef<const gmx::RVec> x,
                                   nbnxn_atomdata_t*              nbat,
                                   const gmx::Range<int>          columnRange,
@@ -1100,7 +1100,7 @@ void Grid::sortColumnsCpuGeometry(GridSetData*                   gridSetData,
 /* Spatially sort the atoms within one grid column */
 void Grid::sortColumnsGpuGeometry(GridSetData*                   gridSetData,
                                   int                            dd_zone,
-                                  gmx::ArrayRef<const int>       atomInfo,
+                                  gmx::ArrayRef<const int64_t>   atomInfo,
                                   gmx::ArrayRef<const gmx::RVec> x,
                                   nbnxn_atomdata_t*              nbat,
                                   const gmx::Range<int>          columnRange,
@@ -1370,7 +1370,7 @@ void Grid::setCellIndices(int                            ddZone,
                           GridSetData*                   gridSetData,
                           gmx::ArrayRef<GridWork>        gridWork,
                           const gmx::Range<int>          atomRange,
-                          gmx::ArrayRef<const int>       atomInfo,
+                          gmx::ArrayRef<const int64_t>   atomInfo,
                           gmx::ArrayRef<const gmx::RVec> x,
                           const int                      numAtomsMoved,
                           nbnxn_atomdata_t*              nbat)
